# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  domain.Admins:
    properties:
      adminid:
        type: integer
      adminname:
        maxLength: 50
        minLength: 2
        type: string
      email:
        type: string
      password:
        type: string
      phonenumber:
        type: string
      verification:
        type: boolean
    required:
    - adminname
    - email
    - password
    type: object
...
  domain.Users:
    properties:
      email:
        type: string
      eventid:
        type: integer
      firstname:
        maxLength: 50
        minLength: 2
        type: string
      lastname:
        maxLength: 50
        minLength: 1
        type: string
      password:
        type: string
      phonenumber:
        type: string
      profile:
        type: string
      userid:
        type: integer
      username:
        maxLength: 50
        minLength: 2
        type: string
      verification:
        type: boolean
      vip:
        type: boolean
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: sethukumarj.76@gmail.com
    name: API Support
    url: sethukumarj.com
  description: This is an Events Radar project. You can visit the GitHub repository
    at https://github.com/SethukumarJ/Events_Radar_Developement
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Radar API
  version: "1.0"
paths:
...
  /admin/login:
    post:
      operationId: Admin Login
      parameters:
      - description: 'adminlogin: '
        in: body
        name: AdminLogin
        required: true
        schema:
          $ref: '#/definitions/domain.Admins'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login for Admin
      tags:
      - Admin
      - Admin
...
  /admin/signup:
    post:
      operationId: SignUp authentication
      parameters:
      - description: admin signup with username, phonenumber email ,password
        in: body
        name: RegisterAdmin
        required: true
        schema:
          $ref: '#/definitions/domain.Admins'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: SignUp for Admin
      tags:
      - Admin
      - Admin
  /admin/token-refresh:
    post:
      operationId: Admin RefreshToken
      parameters:
      - description: 'token string: '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Refresh token for admin
      tags:
      - Admin
      - Admin
...
  /user/login:
    post:
      operationId: User Login
      parameters:
      - description: 'userlogin: '
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/domain.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login for users
      tags:
      - User
      - User
...
  /user/signup:
    post:
      operationId: User SignUp
      parameters:
      - description: user signup with username, phonenumber email ,password
        in: body
        name: RegisterUser
        required: true
        schema:
          $ref: '#/definitions/domain.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: SignUp for users
      tags:
      - User
  /user/token-refresh:
    post:
      operationId: User RefreshToken
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Refresh token for users
      tags:
      - User
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"